Fix to HibernateTreeDataServiceImpl to not attempt to delete items with null ID.  These items are detached and cannot be associated with a row in the DB.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@713 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/treeutils/HibernateTreeDataServiceImpl.java
@@ -163,13 +163,12 @@ public class HibernateTreeDataServiceImpl implements TreeDataService
 		// delete all of the tree def items that were deleted by the user
 		for(TreeDefinitionItemIface item: deletedItems)
 		{
-//			// ignore the items with null ID
-//			// they were probably created, then deleted, before ever being persisted
-//			if(item.getTreeDefItemId() != null)
-//			{
-//				session.delete(item);
-//			}
-			session.delete(item);
+			// ignore the items with null ID
+			// they were probably created, then deleted, before ever being persisted
+			if(item.getTreeDefItemId() != null)
+			{
+				session.delete(item);
+			}
 		}
 		
 		try
